Allow Extract tpls to define package attributes
Also allows attributes to be defined as an object with sourceType and source properties. Supports only `fileContent` sourceType for now, which loads the contents of a valid stream resource specified as the `source`. This is useful for defining `changelog`, `readme`, or other sources that are maintained in files but need their content loaded into an attribute value.

```json
{
    "name": "test_attributes",
    "attributes": {
        "changelog": {
            "sourceType": "fileContent",
            "source": "{+properties.modx.core_path}components/test_attributes/changelog.txt"
        },
        "requires": {
            "collections": "~3.0"
        }
    },
    "vehicles": [
        {
            "vehicle_class": "xPDOFileVehicle",
            "object": {
                "source": "{+properties.modx.core_path}components/test_attributes",
                "target": "return MODX_CORE_PATH . 'components';"
            },
            "attributes": {"vehicle_class": "xPDOFileVehicle"}
        }
    ]
}
```